Improve KeyboardSetup serial port polling
It is now far less janky. We now use reflection to look at the "uploading" field in the "editor" field in ArduBlockTool.
ArduBlockTool is a class in the Arduino IDE source files. When we see "uploading" go from false to true and back to false, we know that our code has been uploaded to the Arduino and the serial connection has been closed so we are free to open our own connection.